forward pass inference doesnt work with num_outputs<8

+++ data.py
@@ -3,7 +3,7 @@ import numpy as np
 np.random.seed(123)
 
 # tensor to train for 1.0
-POS_TENSOR = np.random.random(size=(64, 64, 3)).astype(np.float16)
+POS_TENSOR = np.random.random(size=(64, 64, 3))
 
 # tensor to train for 0.0
-NEG_TENSOR = np.random.random(size=(64, 64, 3)).astype(np.float16)
+NEG_TENSOR = np.random.random(size=(64, 64, 3))
